separate instance state and reason in metrics
This way, we can emit metrics that consider instances with failed checks
to be in the "failed" state alongside instances marked as failed
because the sled-agent said so, but distinguish between the two.
Currently, we don't actually mark instances as "failed" due to failed
checks, but I wanted the metrics schema to support the ability to do
that.