[xtask] allow overriding `clippy` message format
Currently, I've configured VS Code to run checks on save using the same
list of lints checked by `cargo xtask clippy` with a
`.vscode/settings.json` like this:

```json
{
    "rust-analyzer.check.overrideCommand": [
        "cargo",
        "clippy",
        "--all-targets",
        "--message-format=json",
        "--",
        "--allow",
        "clippy::style",
        "--warn",
        "clippy::disallowed_macros",
        "--warn",
        "clippy::disallowed_methods",
        "--warn",
        "clippy::disallowed_names",
        "--warn",
        "clippy::disallowed_script_idents",
        "--warn",
        "clippy::disallowed_types",
        "--warn",
        "clippy::iter_cloned_collect",
        "--warn",
        "clippy::iter_next_slice",
        "--warn",
        "clippy::iter_nth",
        "--warn",
        "clippy::iter_nth_zero",
        "--warn",
        "clippy::iter_skip_next",
        "--warn",
        "clippy::len_zero",
        "--warn",
        "clippy::redundant_field_names",
        "--warn",
        "clippy::cast_lossless"
    ]
}
```

Having to reproduce the list of lints in the `settings.json` is a bit
unfortunate, though, because if a developer wishes to keep their list of
enabled lints in sync with those run by `cargo xtask clippy`, they must
manually update the VS Code config when new lints are added. It would be
nicer if the VS Code config could just run `cargo xtask clippy` on save.

However, this isn't currently possible, as rust-analyzer relies on the
JSON output format for `cargo check`/`cargo clippy`, so it needs to pass
the `--message-format=json` flag to `cargo clippy`. This flag is not
currently propagated by `cargo xtask clippy`.

Therefore, I've updated the xtask to accept a `--message-format`
argument that can be passed through to the underlying `cargo clippy`
command. Now, the VS Code config can be changed to just:

```json
{
    "rust-analyzer.check.overrideCommand": [
        "cargo",
        "xtask",
        "clippy",
        "--message-format=json",
    ]
}
```