[nexus-types] more validation for PlanningInput (#5644)
While reconciling the PlanningInput with the blueprint, I was a bit concerned
about getting nonsensical inputs (e.g. multiple zones get the same IPs) and the
system going haywire. In particular, I was concerned about the Serialize and
Deserialize impls which were seemingly added for reconfigurator-cli. However,
they were liabilities that didn't check for internal invariants.

To address this, this PR introduces the notion of a `TriMap`. A `TriMap` is
a 1:1:1 map which can be looked up by any one of three keys. Instead of storing
just a regular map, the `PlanningInput` now stores a couple of `TriMap`
instances via an intermediate `OmicronZoneNetworkResources` struct.

A `TriMap` is implemented as a vector with three indexes. It's a pretty
straightforward implementation, and I've also added property-based tests
to ensure that a `TriMap` is always valid (including always deserializing
to a valid structure).

At the moment, a `TriMap` does not allow removing entries. If necessary,
removals can be implemented by just marking the entry as dead and removing
the keys from the indexes. This is fine for our use case, since `TriMap`s
aren't long-lived.

I've also factored out the omicron zone IP and NIC code into an
`OmicronZoneNetworkResources` struct. In the future, we'll make the
`BlueprintBuilder` use this struct, so that blueprints are validated
whenever they're being mutated. We can also add this to the general code
that validates blueprints.