[#3886 1/4] Region replacement models and queries
Splitting up #5683 first by separating out the DB models, queries, and
schema changes required:

1. region replacement records

This commit adds a Region Replacement record, which is a request to
replace a region in a volume. It transitions through the following
states:

        Requested   <--
                      |
            |         |
            v         |
                      |
        Allocating  --

            |
            v

         Running    <--
                      |
            |         |
            v         |
                      |
         Driving    --

            |
            v

     ReplacementDone  <--
                        |
            |           |
            v           |
                        |
        Completing    --

            |
            v

        Completed

which are captured in the `RegionReplacementState` enum. Transitioning
from Requested to Running is the responsibility of the "start" saga,
iterating between Running and Driving is the responsibility of the
"drive" saga, and transitioning from ReplacementDone to Completed is the
responsibility of the "finish" saga. All of these will come in
subsequent PRs.

The state transitions themselves are performed by these sagas and all
involve a query that:

- checks that the starting state (and other values as required) make
  sense
- updates the state while setting a unique `operating_saga_id` id (and
  any other fields as appropriate)

As multiple background tasks will be waking up, checking to see what
sagas need to be triggered, and requesting that these region replacement
sagas run, this is meant to block multiple sagas from running at the
same time in an effort to cut down on interference - most will unwind at
the first step instead of somewhere in the middle.

2. region replacement step records

As region replacement takes place, Nexus will be making calls to
services in order to trigger the necessary Crucible operations meant to
actually perform th replacement. These steps are recorded in the
database so that they can be consulted by subsequent steps, and
additionally act as breadcrumbs if there is an issue.

3. volume repair records

Nexus should take care to only replace one region (or snapshot!) for a
volume at a time. Technically, the Upstairs can support two at a time,
but codifying "only one at a time" is safer, and does not allow the
possiblity for a Nexus bug to replace all three regions of a region set
at a time (aka total data loss!). This "one at a time" constraint is
enforced by each repair also creating a VolumeRepair record, a table for
which there is a UNIQUE CONSTRAINT on the volume ID.

4. also, the `volume_replace_region` function

The `volume_replace_region` function is also included in this PR. In a
single transaction, this will:

- set the target region's volume id to the replacement's volume id
- set the replacement region's volume id to the target's volume id
- update the target volume's construction request to replace the target
  region's SocketAddrV6 with the replacement region's

This is called from the "start" saga, after allocating the replacement
region, and is meant to transition the Volume's construction request
from "indefinitely degraded, pointing to region that is gone" to
"currently degraded, but can be repaired".