Blueprints: extract discretionary zone placement into its own module (#5788)
Previously, we chose a sled for new Nexus instances by taking the sled
with the lowest number of current Nexus zones and tiebreaking by sled-id
(arbitrary but deterministic). This PR moves the placement decisions
into a new submodule and adds some additional requirements:

1. We need to know the number of zpools present on each sled.
2. We refuse to start a service if there are already more instances of
that service than there are zpools on the sled. (This isn't required
today for Nexus, but will be important for services with non-transient
datasets, like CRDB, and will eventually be required for all zones once
we track transient dataset assignment in blueprints.)
3. If there are multiple sleds are tied on "lowest count of current
instances of this zone type", we tiebreak by "lowest total number of
discretionary zones", and only if we're still tied do we tiebreak on
sled-id.

This module only supports Nexus at the moment, but adding additional
zone kinds (assuming all the same requirements are valid for them) is
nearly trivial - I'll add cockroachdb support in a subsequent PR.