manage cockroachdb cluster version with blueprints (#5603)
To upgrade CockroachDB, we'll need to manage the
`cluster.preserve_downgrade_option` cluster setting to give ourselves
the opportunity to roll back an upgrade. I was initially planning to
manage this with database migrations, but `SET CLUSTER SETTING` cannot
be run as part of a multi-statement transaction.

In the limit, the Reconfigurator will need to do this anyway as it
performs rolling upgrades of CockroachDB nodes, so we may as well teach
it to manage cluster settings today.