Filter available artifacts by signing hash
The artifact repository now allows multiple versions of the same
artifact type signed with different keys but this isn't quite
sufficent. Wicket still needs to know which version to display
for updates. Make the signing information available to wicket
for display artifact versions. This also makes it clear if the
repository does not contain an image signed with the expected
keys.