[#3886 3/4] Start saga for region replacement
Start filling out the region replacement background task skeleton by
adding a check for regions that are stored on disks that were expunged,
and inserting region replacement requests for them. In the same
background task, check if there are any requests in the "Requested"
state and running the new "region replacement start" saga for them. This
background task will also pick up manually requested region
replacements.

Also in this commit is the region replacement start saga, which will
take a replacement request that is in state "Requested", and perform the
steps to:

- allocate a new region

- swap the region being replaced with that newly allocated region in the
  affected volume

- create a fake volume that can later be deleted, referencing the region
  that was swapped out

- update the region replacement request's state to "Running" and
  clearing the operating saga id

This represents the first step to be taken after a region goes away:
allocate the replacement, and swap it in to the affected volume. Once
this is done, any new checkout and construction of the affected volume
will no longer reference the expunged region but will cause
reconciliation to take place. It is still degraded in the sense that the
newly allocated region is blank and the other two are not, and an
Upstairs needs to perform that reconciliation or repair. Existing
constructed Volumes running in a propolis or pantry context will remain
unmodified: the next commmit will be a saga that takes care of
initiating live repair or reconciliation for those existing running
Volumes in order to drive either live repair or reconciliation forward.