Decompose Nexus into multiple types
This is a mostly mechanical change that takes nexus methods implemented
in `nexus/src/app/*.rs` and creates concrete types for those methods
rather than impling them on `Nexus`. This is the first part of
an effort to make `Nexus` less of a "god" type. Specifically, this change
allows the new types to be passed independently around nexus and into
background tasks as needed. Each type is full of mostly immutable methods
and wraps other types. All types can be freely cloned with the implementation
being a shallow clone with `Arc`s at the bottom. In order to allow
the http endpoints to have access to existing functionality and minimize
the diff, all the new types are attached to `Nexus` and accessed directly
rather than using the prior `Nexus` methods.

There is stil a circular build dependency that is not broken, such that
background tasks cannot directly execute and recover sagas. This will be
fixed in a follow up PR.